Add the previous message to GPT-3 query
GPT-3 now does not receive only the new message, but also the previous one. Either the message replied to or the preceding message in the same chat is sent as well. This will allow for taking some conversation context into account.

For this, the whole prompt has been rewritten to resemble a chat conversation.