Use GPT with vision if there are images in chat
Previously, OpenAI only had a model with vision that didn’t support function calling. Thus vision was implemented as an extra tool to call, while the response was made with an extra GPT call. Now the GPT-4 Turbo model supports vision and function calling, enabling a direct implementation.

The bot will use GPT-3.5 Turbo by default if there are no images in the past 15 messages it sees, but use GPT-4 Turbo if there is at least one to enable vision.