Add wording to the Stats tool to call the Search tool for card info
The bot sometimes uses the Stats tool when the user asks for missing or owned cards when it should actually use the Search tool. The Stats tool will now tell the bot to call the Search tool if the user asked for specific cards.

This will hopefully make the bot to call the correct tool too.

no_announcement