Store tool calls in database
Currently, tool calls are not stored. This makes the bot “forget” that it actually called tools in between messages when generating a new message from chat history.

This change now stores tool calls to help the bot to understand the actual chat history, especially useful if later redoing the command call to know that its knowledge came from a tool and not from invention.